Move wheel to operations.install.wheel
